android-build: add required files to build the sample from Android sources
This sample will be included as a default application in the Digi Android reference
images. In order to build the sample from the Android sources, a special Android build
script is required, as well as a permissions file. This commit adds these files to the
project so the sample can be built as is after clonning it from the Android sources.

https://onedigi.atlassian.net/browse/ADK4A-1850

Signed-off-by: David Escalona <david.escalona@digi.com>