android.bp: disable `<uses-library>` check in `AndroidManifest.xml`
To construct the correct class loader context for dexpreopt, the build system
needs to be aware of the shared library dependencies of Java modules declared in
the `<uses-library>` tags in the manifest. Since the build system does not have
direct access to the manifest, this information must be provided in the build
files.

In our case, the manifest uses `com.digi.android` library. When building from
sources, Soong can infer `uses_libs` because it has knowledge of Java SDK
libraries and the `libs` property in `Android.bp`. However, when building
against pre-built binaries, this inference does not occur, and the build fails
with the following error:

error: mismatch in the <uses-library> tags between the build system and the manifest:
	- required libraries in build system: []
	                 vs. in the manifest: [com.digi.android]
	- optional libraries in build system: []
	                 vs. in the manifest: []
	- tags in the manifest (out/soong/.intermediates/vendor/digi/firmwareupdatesample/FirmwareUpdateSample/android_common/737f09fcc0b17bc4650c68e8a26b3dd4/manifest_merger/AndroidManifest.xml):
		<uses-library android:name="com.digi.android" android:required="true"/>

note: the following options are available:
	- to temporarily disable the check on command line, rebuild with RELAX_USES_LIBRARY_CHECK=true (this will set compiler filter "verify" and disable AOT-compilation in dexpreopt)
	- to temporarily disable the check for the whole product, set PRODUCT_BROKEN_VERIFY_USES_LIBRARIES := true in the product makefiles
	- to fix the check, make build system properties coherent with the manifest
	- for details, see build/make/Changes.md and https://source.android.com/devices/tech/dalvik/art-class-loader-context

This commit disables the check by setting `enforce_uses_libs` to `false`.

https://onedigi.atlassian.net/browse/CCS-29

Signed-off-by: Tatiana Leon <Tatiana.Leon@digi.com>